# backend/src/main/java/com/vobworkbench/feature/patient/dto/CreatePatientRequest.java
package com.vobworkbench.feature.patient.dto;

import com.vobworkbench.feature.patient.entity.Gender;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Past;

import java.time.LocalDate;

public record CreatePatientRequest(

        @NotBlank String mrn,
        @NotBlank String firstName,
        @NotBlank String lastName,
        @NotNull @Past LocalDate dateOfBirth,
        @NotNull Gender gender,
        @NotBlank String phone
) {
}