# docs/openapi.yaml
openapi: 3.0.3
info:
  title: VOB Workbench API
  version: 0.0.1
  description: Authentication and patient intake API for the VOB Workbench backend.
servers:
  - url: http://localhost:8080
    description: Local backend
...
paths:
...
  /api/patients:
    post:
      tags:
        - Patients
      summary: Create a patient during front desk intake.
      operationId: createPatient
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePatientRequest'
            example:
              mrn: MRN-10001
              firstName: Jane
              lastName: Doe
              dateOfBirth: '1990-04-10'
              gender: FEMALE
              phone: '555-123-4567'
      responses:
        '201':
          description: Patient created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Authenticated user lacks PATIENT_CREATE.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Patient MRN already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    get:
      tags:
        - Patients
      summary: List patients using cursor-based pagination.
      operationId: listPatients
      security:
        - bearerAuth: []
      parameters:
        - name: cursor
          in: query
          required: false
          description: Opaque cursor returned from the previous page.
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: Number of records to return.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Patient page returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientPageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Authenticated user lacks PATIENT_VIEW.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/patients/{id}:
    get:
      tags:
        - Patients
      summary: Get a patient by id.
      operationId: getPatientById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: 6682df563cc7f568fbf16210
      responses:
        '200':
          description: Patient returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Authenticated user lacks PATIENT_VIEW.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Patient not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Missing or invalid JWT.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
...
    CreatePatientRequest:
      type: object
      required:
        - mrn
        - firstName
        - lastName
        - dateOfBirth
        - gender
        - phone
      properties:
        mrn:
          type: string
          minLength: 1
          example: MRN-10001
        firstName:
          type: string
          minLength: 1
          example: Jane
        lastName:
          type: string
          minLength: 1
          example: Doe
        dateOfBirth:
          type: string
          format: date
          example: '1990-04-10'
        gender:
          $ref: '#/components/schemas/Gender'
        phone:
          type: string
          minLength: 1
          example: '555-123-4567'
    PatientResponse:
      type: object
      required:
        - id
        - mrn
        - firstName
        - lastName
        - dateOfBirth
        - gender
        - phone
        - createdByUserId
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          example: 6682df563cc7f568fbf16210
        mrn:
          type: string
          example: MRN-10001
        firstName:
          type: string
          example: Jane
        lastName:
          type: string
          example: Doe
        dateOfBirth:
          type: string
          format: date
          example: '1990-04-10'
        gender:
          $ref: '#/components/schemas/Gender'
        phone:
          type: string
          example: '555-123-4567'
        createdByUserId:
          type: string
          example: 6682df563cc7f568fbf11111
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PatientPageResponse:
      type: object
      required:
        - items
        - hasNext
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PatientResponse'
        nextCursor:
          type: string
          nullable: true
          description: Opaque cursor for the next page. Built from createdAt and id, but clients should not parse it.
        hasNext:
          type: boolean
          example: true
...
    Gender:
      type: string
      enum:
        - MALE
        - FEMALE
        - OTHER
        - UNKNOWN
    ApiError:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Request validation failed
        path:
          type: string
          example: /api/patients